==> scbcr_seq/02_celltype_annotation.Rmd <==
---
title: "02_celltype_annotation"
author: ""
date: "2024-01-03"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```

## Load library
```{r}
library(Seurat)
library(SeuratData)
library(future)
library(Matrix)
```


## Data
```{r}
plan(strategy = "multicore", workers = 30)
seu.all <- readRDS("../results/All_SeuratObject_annotated.RDS")

count.list <- list(counts.1=seu.all@assays$RNA$counts.1,
                   counts.2=seu.all@assays$RNA$counts.2,
                   counts.3=seu.all@assays$RNA$counts.3,
                   counts.4=seu.all@assays$RNA$counts.4,
                   counts.5=seu.all@assays$RNA$counts.5)
data.list <- list(data.1=seu.all@assays$RNA$data.1,
                  data.2=seu.all@assays$RNA$data.2,
                  data.3=seu.all@assays$RNA$data.3,
                  data.4=seu.all@assays$RNA$data.4,
                  data.5=seu.all@assays$RNA$data.5)
count.mat <- Reduce(RowMergeSparseMatrices, count.list)
data.mat <- Reduce(RowMergeSparseMatrices, data.list)

seu.all <- SetAssayData(object = seu.all, assay = "RNA", slot = "counts", new.data = count.mat)
seu.all <- SetAssayData(object = seu.all, assay = "RNA", slot = "data", new.data = data.mat)

# plan(strategy = "multicore", workers = 30)
# pos.markers <- FindAllMarkers(seu.all, only.pos = TRUE)
# write.csv(pos.markers, file="../results/Cluster_pos_markers_res1.csv")
```


## Annotation
```{r}
pos.markers <- read.csv("../results/Cluster_pos_markers_res1.csv")

top10 <- pos.markers %>%
  group_by(cluster) %>%
  dplyr::filter(avg_log2FC > 1) %>%
    slice_head(n = 10) %>%
    ungroup()
  
DoHeatmap(seu.all, features = top10$gene) + NoLegend()




library(RColorBrewer)
set.seed(123)
n <- 40
qual_col_pals = brewer.pal.info[brewer.pal.info$category == 'qual',]
col_vector = unlist(mapply(brewer.pal, qual_col_pals$maxcolors, rownames(qual_col_pals)))

DimPlot(seu.all, reduction = "umap", cols = col_vector,  group.by = "seurat_clusters", label = TRUE, label.size = 5, repel = TRUE)


DimPlot(seu.all, reduction = "umap", cols = col_vector,  group.by = "seurat_clusters", label = TRUE, label.size = 5, repel = TRUE, split.by = "Group")
```


```{r}
library(EasyCellType)
library(org.Mm.eg.db)
library(AnnotationDbi)
pos.markers$entrezid <- mapIds(org.Mm.eg.db,
                           keys=pos.markers$gene, #Column containing Ensembl gene ids
                           column="ENTREZID",
                           keytype="SYMBOL",
                           multiVals="first")
pos.markers <- na.omit(pos.markers)


library(dplyr)
markers_sort <- data.frame(gene=pos.markers$entrezid, cluster=pos.markers$cluster, 
                      score=pos.markers$avg_log2FC) %>% 
  group_by(cluster) %>% 
  mutate(rank = rank(score),  ties.method = "random") %>% 
  arrange(desc(rank)) 
input.d <- as.data.frame(markers_sort[, 1:3])


annot.GSEA <- easyct(input.d, db="cellmarker", species="Mouse", p_cut=0.3,
                    test="GSEA")
plot_dot(test="GSEA", annot.GSEA)
plot_bar(test="GSEA", annot.GSEA)

```



```{r}
top20 <- pos.markers %>%
  group_by(cluster) %>%
  dplyr::filter(avg_log2FC > 0.5) %>%
    slice_head(n = 10) %>%
    ungroup()
clus <- unique(top20$cluster)
f.list <- list()
for (i in 1:length(clus)) {
  f.list[i] <- paste0("cluster_", clus[i], ":", paste0(top20$gene[top20$cluster == clus[i]], collapse = ","))
}
f.df <- do.call(rbind, f.list)
write.table(f.df, file = "cluster_feature.tsv", sep = "\t")
```
